+++ app/src/components/admin/index.ts
@@ -4,6 +4,7 @@
 
 export { AdminAuthProvider, useAdminAuth } from "./AdminAuthProvider";
 export { AdminLayout } from "./AdminLayout";
+export { AdminLogo } from "./AdminLogo";
 export { AdminPageContent } from "./AdminPageContent";
 export { AdminProvider } from "./AdminProvider";
 export { AdminThemeContext, useAdminTheme } from "./admin-theme";