+++ app/src/components/client/index.ts
@@ -4,5 +4,6 @@
 
 export { CaptchaInput } from "./CaptchaInput";
 export { ClientAuthProvider, useClientAuth } from "./ClientAuthProvider";
+export { ClientLogo } from "./ClientLogo";
 export { Footer } from "./Footer";
 export { Header } from "./Header";